Refactor layout transfer validation into compiler core
